Changed REVOLVING to ORBITING and those circles orbit at different rates
ORIBTING semeantically makes more sense and having circles orbit at
different rates was one of my early goals which also adds flair to the
animation.

Additionally, removed Google fonts import so that it relies on the
browser default. I may change this in the future to a font to ensure
that weird representations don't occur through CharacterMapper
processing.